// web/src/form/questions/response/__generated__/LinearScaleQuestionResponseQuery.graphql.ts
export type LinearScaleQuestionResponseQuery$variables = {
  formInstanceID: string;
  questionID: string;
};
export type LinearScaleQuestionResponseQuery$data = {
  readonly questionResponses: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly __typename: "QuestionResponse";
        readonly id: string;
        readonly value: string;
      } | null | undefined;
    } | null | undefined> | null | undefined;
  };
};
export type LinearScaleQuestionResponseQuery = {
  response: LinearScaleQuestionResponseQuery$data;
  variables: LinearScaleQuestionResponseQuery$variables;
};